refactor of db functions
